llmops-code-genration-sample/llmops-evaluation/requirements.txt
langfuse>=2.0.0
mlflow>=2.10.0
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0
opentelemetry-exporter-otlp>=1.21.0
opentelemetry-instrumentation>=0.42b0
python-dotenv>=1.0.0
requests>=2.31.0